Seperate config package
